refactor: Update Results.vue to display authors of library items
The Results.vue component has been updated to display the authors of library items in the search results. This change adds a new section that shows the names of the authors for each item, separated by commas. The code has been modified to iterate over the authors array and render the names accordingly.

Refactor the Results.vue component to improve the display of library item authors.